requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
openai==1.3.0
pymongo==4.6.0
motor==3.3.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.4.2
openai-whisper==20231117
torch==2.1.0
numpy==1.24.3
Jinja2==3.1.2 